+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-* v0.1.0
+## v0.1.1
+* Fix missing queue message in `after_commit on: :destroy`
 
-- No longer queue/delay in after_save/after_commit if no changes occurred
-- Added Redstream.stream_size
+## v0.1.0
+* No longer queue/delay in `after_save`/`after_commit` if no changes occurred
+* Added `Redstream.stream_size`